fixate, completed authorize endpoint (OAuth2.1)

+++ internal/domain/models/authorization_code.go
@@ -8,6 +8,6 @@ import (
 // AuthorizationCode model depends on RFC OAUTH2.1
 type AuthorizationCode struct {
 	Code      uu.UUID   `json:"code" db:"code"`
-	UserID    int       `json:"user_id" db:"user_id"`
+	UserID    int64     `json:"user_id" db:"user_id"`
 	ExpiresAt time.Time `json:"expires_at" db:"expires_at"`
 }
